+++ Patches/PatchTerminal.cs
@@ -15,12 +15,12 @@ public class PatchTerminal
     {
         string[] array = __instance.screenText.text.Substring(__instance.screenText.text.Length - __instance.textAdded).Split(' ', System.StringSplitOptions.RemoveEmptyEntries);
         
-        if (array[0].ToLower().Contains("reorder"))
+        if (array[0].ToLower().Contains("reorder") || array[0].ToLower().Contains("reoder"))
         {
 
             if (array.Length > 1 && array[1].ToLower().Contains("help"))
             {
-                __result = CreateTerminalNode("Commands :\n\n- reorder\n\n- reorder <valueRange>\n\n- reorder name\n\n");
+                __result = CreateTerminalNode("Commands :\n\n- reorder\n\n- reorder <valueRange>\n\n- reorder name\n\n- reorder type\n\n");
                 return false;
             }
 
@@ -28,6 +28,7 @@ public class PatchTerminal
             var value = OptimalScrapsOrganizationPlugin.instance.organiseDefaultValueRange.Value;
             
             if(array.Length > 1 && array[1].ToLower().Contains("name")) organizeBy = OrganizeBy.NAME;
+            else if(array.Length > 1 && array[1].ToLower().Contains("type")) organizeBy = OrganizeBy.TYPE;
             else if (array.Length > 1)
             {
                 organizeBy = OrganizeBy.VALUE;
@@ -37,7 +38,11 @@ public class PatchTerminal
             OrganizeInformation organizeInformation = new OrganizeInformation();
             organizeInformation.OrganizeBy = organizeBy;
             organizeInformation.value = value;
-            organizeInformation.distanceBetweenObjects = OptimalScrapsOrganizationPlugin.instance.distanceBewteenObjects.Value;
+            organizeInformation.distanceBetweenObjects = OptimalScrapsOrganizationPlugin.instance.distanceBetweenScraps.Value;
+            organizeInformation.rotateScraps = OptimalScrapsOrganizationPlugin.instance.rotateScraps.Value;
+            organizeInformation.rotationBetweenScraps = OptimalScrapsOrganizationPlugin.instance.rotationBetweenScraps.Value;
+            organizeInformation.exclusionList = OptimalScrapsOrganizationPlugin.instance.exclusionList.Value;
+            organizeInformation.orderShopItems = OptimalScrapsOrganizationPlugin.instance.orderShopItems.Value;
             
             NetworkOrganization.OrganizeScrapsServerRpc(organizeInformation);
             __result = CreateTerminalNode("Done !");
@@ -51,6 +56,9 @@ public class PatchTerminal
     [HarmonyPriority(Priority.Last)]
     private static void StartPatch(ref TerminalNodesList ___terminalNodes)
     {
+        
+        if(OptimalScrapsOrganizationPlugin.instance.hasPatchedStartTerminal) return;
+        
         int index = 1;
         string defaultMessage = ___terminalNodes.specialNodes[index].displayText;
 
@@ -66,6 +74,9 @@ public class PatchTerminal
             $"\n\n>REORDER HELP\nTo see the list of {OptimalScrapsOrganizationPlugin.NAME} commands.\n";
 
         ___terminalNodes.specialNodes[index].displayText = message;
+
+        OptimalScrapsOrganizationPlugin.instance.hasPatchedStartTerminal = true;
+
     }
     
     public static TerminalNode CreateTerminalNode(string message, bool clearPreviousText = true, int maxChar = 50)