+++ Scripts/OrganizeInformation.cs
@@ -5,4 +5,8 @@ public class OrganizeInformation
     public OrganizeBy OrganizeBy;
     public int value;
     public float distanceBetweenObjects;
+    public bool rotateScraps;
+    public float rotationBetweenScraps;
+    public string exclusionList;
+    public bool orderShopItems;
 }