+++ Scripts/ScrapsOrganizationManager.cs
@@ -16,12 +16,12 @@ public class ScrapsOrganizationManager
     public static Vector3 postionRight = new Vector3(-4.263417f,0,-9.5f);
     public static Vector3 maxPostionRight = new Vector3(3.728537f,0,-9.5f);
     
-    public static bool IsScrap(GrabbableObject grabbableObject)
+    public static bool IsScrap(GrabbableObject grabbableObject, OrganizeInformation organizeInformation)
     {
         if (grabbableObject == null) return false;
-        if (!IsScrap(grabbableObject.itemProperties)) return false;
+        if (!IsScrap(grabbableObject.itemProperties, organizeInformation)) return false;
 
-        if (grabbableObject.isHeld || !grabbableObject.grabbable)
+        if (grabbableObject.isHeld || !grabbableObject.grabbable || IsExcluded(grabbableObject, organizeInformation))
         {
             return false;
         }
@@ -29,39 +29,56 @@ public class ScrapsOrganizationManager
         return true;
     }
 
-    public static bool IsScrap(Item item)
+    public static bool IsExcluded(GrabbableObject grabbableObject, OrganizeInformation organizeInformation)
+    {
+        var result = false;
+        if(organizeInformation.exclusionList == "") return false;
+        
+        organizeInformation.exclusionList.Split(",").ToList().ForEach(name =>
+        {
+            if(organizeInformation.exclusionList == "")
+            {
+                result = false;
+            }
+            else if(OptimalScrapsOrganizationPlugin.instance.StringContain(grabbableObject.itemProperties.itemName, name.ToString())) result = true;
+        });
+
+        return result;
+    }
+
+    public static bool IsScrap(Item item, OrganizeInformation organizeInformation)
     {
         if (item == null) return false;
 
-        return item.isScrap;
+        return organizeInformation.orderShopItems || item.isScrap;
     }
     
-    public static bool IsValidScrap(GrabbableObject grabbableObject)
+    public static bool IsValidScrap(GrabbableObject grabbableObject, OrganizeInformation organizeInformation)
     {
         if (grabbableObject == null) return false;
 
-        if (!IsScrap(grabbableObject)) return false;
+        if (!IsScrap(grabbableObject, organizeInformation)) return false;
         
         return true;
     }
     
-    private static List<GrabbableObject> GetValidScrap(IEnumerable<GrabbableObject> grabbableObjects)
+    private static List<GrabbableObject> GetValidScrap(IEnumerable<GrabbableObject> grabbableObjects, OrganizeInformation organizeInformation)
     {
         if (grabbableObjects == null) return [];
 
-        return grabbableObjects.Where(x => IsValidScrap(x)).ToList();
+        return grabbableObjects.Where(x => IsValidScrap(x, organizeInformation)).ToList();
     }
     
-    public static List<GrabbableObject> GetScrapFromShip()
+    public static List<GrabbableObject> GetScrapFromShip(OrganizeInformation organizeInformation)
     {
         if (HangarShipTransform == null) return [];
 
-        return GetValidScrap(HangarShipTransform.GetComponentsInChildren<GrabbableObject>());
+        return GetValidScrap(HangarShipTransform.GetComponentsInChildren<GrabbableObject>(), organizeInformation);
     }
 
     public static void OrganizeShipScraps(OrganizeInformation organizeInformation)
     {
-        var scrapsOnShip = GetScrapFromShip();
+        var scrapsOnShip = GetScrapFromShip(organizeInformation);
 
         var value = organizeInformation.value;
         var organizeBy = organizeInformation.OrganizeBy;
@@ -80,6 +97,11 @@ public class ScrapsOrganizationManager
                 scrapsOnShip.Sort((x, y) => x.itemProperties.itemName.CompareTo(y.itemProperties.itemName));
                 break;
             }
+            case OrganizeBy.TYPE:
+            {
+                scrapsOnShip.Sort((x, y) => x.itemProperties.twoHanded.CompareTo(y.itemProperties.twoHanded));
+                break;
+            }
         }
         
         if(scrapsOnShip.Count == 0) return;
@@ -87,6 +109,7 @@ public class ScrapsOrganizationManager
         var basePos = position;
         var usePostion = basePos;
         var valueRange = 0;
+        var rotation = 0f;
 
         while (valueRange < scrapsOnShip[0].scrapValue)
         {
@@ -112,9 +135,15 @@ public class ScrapsOrganizationManager
             
             scrap.targetFloorPosition = usePostion;
             scrap.transform.eulerAngles = Vector3.zero + scrap.itemProperties.restingRotation;
+            if (organizeInformation.rotateScraps)
+            {
+                scrap.transform.eulerAngles += new Vector3(0, rotation, 0);
+                rotation += organizeInformation.rotationBetweenScraps;
+            }
 
             if (organizeBy == OrganizeBy.NAME && index < scrapsOnShip.Count - 1 &&  scrap.itemProperties.itemName == scrapsOnShip[index+1].itemProperties.itemName) return;
             if (organizeBy == OrganizeBy.VALUE && index < scrapsOnShip.Count - 1  && scrapsOnShip[index + 1].scrapValue <= valueRange ) return;
+            if (organizeBy == OrganizeBy.TYPE && index < scrapsOnShip.Count - 1  && scrapsOnShip[index + 1].itemProperties.twoHanded == scrap.itemProperties.twoHanded ) return;
 
             if (index < scrapsOnShip.Count - 1 )
             {
@@ -125,6 +154,7 @@ public class ScrapsOrganizationManager
             }
 
             usePostion += new Vector3(organizeInformation.distanceBetweenObjects, 0, 0);
+            rotation = 0;
 
         });