+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### v1.0.2
+
+- Fix help command text added each time a game is started
+- (Config enabled by default) Added rotation for each object in a stack to make them easier to see
+- (Config enabled by default) Reorder shop items (Jetpack, shovel, flashlight...)
+- Add a ExclusionList config, you can use it so items are not affected by the mod. Example : Shovel,Bell
+- Add a new command "reorder type" to reorder scraps by types (one handed, two handed)
+
 ### v1.0.1
 
 - Add autoReorderOnLeftMoon config, so you don't have to type command each day